+++ package/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 ### What's Changed
 
-- Add default constructor for `RequestBodyLimit`
-- Add default constructor for `RequestTimeLimit`
+- Add missing default constructor for `RequestBodyLimit`
+- Add missing default constructor for `RequestTimeLimit`
 
 ## 0.6.0 (2025-05-20)